demos/dpt/requirements.txt
gradio
torch
torchvision
numpy
pillow
pymatting
opencv-python
onnxruntime-gpu